feat: add dark and light mode to redux
