+++ src/components/Header/Header.tsx
@@ -18,6 +18,11 @@ import { Switch } from "~/components/ui/switch";
 
 import { Sun, Moon } from "lucide-react";
 
+import { useDispatch, useSelector } from 'react-redux';
+import { setIsDay, AppDispatch, RootState } from '../../rtk';
+
+import { MouseEvent } from 'react';
+
 const pages = [
   { id: 1, pathname: "Home" },
   { id: 2, pathname: "Login" },
@@ -25,7 +30,17 @@ const pages = [
 ] as const;
 
 export const Header = () => {
+
   const clickHandler = () => document.body.classList.toggle('dark');
+
+  const dispatch = useDispatch<AppDispatch>();
+  const { isDay } = useSelector((state: RootState) => state.data);
+
+  const handleClick = () => {
+    const isDark = document.body.classList.contains('dark') as boolean;
+    dispatch(setIsDay(isDark));
+  };
+
   return (
     <header className="px-2 pt-2 flex justify-between">
       <NavigationMenu className="grow-0">
@@ -50,7 +65,7 @@ export const Header = () => {
       <div className="flex flex-nowrap items-center">
         <Sun className="mx-2"/>
         <div className="mx-2">
-          <Switch onClick={clickHandler} className="mt-2"/>
+          <Switch onClick={handleClick} className="mt-2"/>
         </div>
         <Moon className="ml-2 mr-4"/>
         <Button>Sign Out</Button>
