+++ src/pages/_app.tsx
@@ -5,16 +5,21 @@ import { SessionProvider } from "next-auth/react";
 import { Layout } from "~/components/Layout";
 import "~/styles/globals.css";
 
+import { Provider } from 'react-redux';
+import { store } from '../rtk';
+
 const MyApp: AppType<{ session: Session | null }> = ({
   Component,
   pageProps: { session, ...pageProps },
 }) => {
   return (
-    <SessionProvider session={session}>
-      <Layout>
-        <Component {...pageProps} />
-      </Layout>
-    </SessionProvider>
+    <Provider store={store}>
+      <SessionProvider session={session}>
+        <Layout>
+          <Component {...pageProps} />
+        </Layout>
+      </SessionProvider>
+    </Provider>
   );
 };
